requirements.txt
folium==0.15.0
branca==0.6.0
python-dotenv==0.21.1
black==23.3.0
geopandas==0.10.2
leafmap==0.23.4
openai==1.3.6
langchain==0.0.344
seaborn==0.13.0